requirements.txt
uvicorn[standard]==0.13.3
fastapi==0.63.0
pydantic==1.7.3
opentelemetry-instrumentation-fastapi==0.25b2
scipy==1.6.0
numpy==1.19.5
sentence-transformers==0.4.1.2
starlette-prometheus==0.9.0
asgiref==3.4.1